src/main/java/com/example/springdocker/service/FoodService.java
package com.example.springdocker.service;

import com.example.springdocker.model.Food;
import com.example.springdocker.repository.FoodRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.stream.Collectors;

@RequiredArgsConstructor
@Service
public class FoodService {
    private final FoodRepository repository;

    public List<Food> getFoods() {
        return repository.findAll();
    }

    public boolean saveNewFood(Food food) {
        List<Food> ids = repository.findFoodById(food.getId());
        boolean taken = false;
        for(Food f : ids){
            if(f.getId().equals(food.getId())){
                taken = true;
            }
        }
        if(taken){
            return false;
        }

        repository.save(food);
        return true;
    }

    public List<String> getCookableFoods() {
        // hämtar alla Foods som vi kan laga
        List<Food> cookableFoods = repository.findFoodByCanICookIt(true);

        // returnerar endast Food namnen i en lista
        return cookableFoods.stream()
                .map(food -> food.getName())
                .collect(Collectors.toList());
    }
}
